Removed CooldownFutureFactory, moved its content to CooldownFuture
